fixing warning and unittest build, refs #3
git-svn-id: file:///home/behr_mi/git/sumo_synched/trunk@12426 afbd958f-9f77-42d5-a016-97a22340ccf4

+++ sumo/src/microsim/devices/MSDevice_Person.h
@@ -119,7 +119,7 @@ public:
      * @return The number of stored persons
      */
     unsigned int size() const {
-        return myPersons.size();
+        return static_cast<unsigned int>(myPersons.size());
     }
 
     const std::vector<MSPerson*> &getPersons() const {
